requirements.txt
pandas 
numpy
setuptools
google-cloud-storage
pyyaml
tensorflow
comet-ml
dvc



